+++ tests/html/CompoundTest.php
@@ -25,6 +25,20 @@ final class CompoundTest extends \PHPUnit\Framework\TestCase
             },
             {
                 "attributes":{
+                    "bold":true,
+                    "script":"sub"
+                },
+                "insert":"quam sapien "
+            },
+            {
+                "attributes":{
+                    "bold":true,
+                    "underline":true
+                },
+                "insert":"quam sapien "
+            },
+            {
+                "attributes":{
                     "strike":true
                 },
                 "insert":"sollicitudin"
@@ -82,7 +96,7 @@ final class CompoundTest extends \PHPUnit\Framework\TestCase
         ]
     }';
 
-    private $expected_multiple_attributes = "<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed efficitur nibh tempor augue lobortis, nec eleifend velit venenatis. Nullam fringilla dui eget lectus mattis tincidunt. Donec sollicitudin, lacus sed luctus ultricies, <s><em>quam sapien </em></s><s>sollicitudin</s> quam, nec auctor eros felis elementum quam. Fusce vel mollis enim. <strong>Sed ac augue tincidunt,</strong> cursus urna a, tempus ipsum. Donec pretium fermentum erat a <u>elementum</u>. In est odio, mattis sed dignissim sed, porta ac nisl. Nunc et tellus imperdiet turpis placerat tristique nec quis justo. Aenean nisi libero, auctor a laoreet sed, fermentum vel massa. Etiam ultricies leo eget purus tempor dapibus. Integer ac sapien eros. Suspendisse convallis ex.</p>";
+    private $expected_multiple_attributes = "<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed efficitur nibh tempor augue lobortis, nec eleifend velit venenatis. Nullam fringilla dui eget lectus mattis tincidunt. Donec sollicitudin, lacus sed luctus ultricies, <s><em>quam sapien </em></s><strong><sub>quam sapien </sub></strong><strong><u>quam sapien </u></strong><s>sollicitudin</s> quam, nec auctor eros felis elementum quam. Fusce vel mollis enim. <strong>Sed ac augue tincidunt,</strong> cursus urna a, tempus ipsum. Donec pretium fermentum erat a <u>elementum</u>. In est odio, mattis sed dignissim sed, porta ac nisl. Nunc et tellus imperdiet turpis placerat tristique nec quis justo. Aenean nisi libero, auctor a laoreet sed, fermentum vel massa. Etiam ultricies leo eget purus tempor dapibus. Integer ac sapien eros. Suspendisse convallis ex.</p>";
     private $expected_multiple_unknown_attributes_image = '<p>Text 1 <strong>assumenda</strong> Text 2.</p><p><img src="data:image/png;base64,ImageDataOmittedforSize" width="214" style="display: inline; float: right; margin: 0px 0px 1em 1em;" /></p><p>Text 3.</p>';
 
     /**
@@ -106,7 +120,7 @@ final class CompoundTest extends \PHPUnit\Framework\TestCase
     }
 
     /**
-     * Test a delta with multiple unknown attributes ona an image, attributes should be included as is
+     * Test a delta with multiple unknown attributes on an image, attributes should be included as is
      *
      * @return void
      * @throws \Exception